requirements-test.txt
pytest>=8.0
pytest-cov>=5.0
pytest-mock>=3.14
pytest-asyncio>=0.24
httpx>=0.27
ruff>=0.4
mypy>=1.10
pyyaml>=6.0
gitpython>=3.1
fastapi>=0.110
